More effective workaround for an environment bug
With this patch, it is less likely for user defined
variables to clash with internal trxl stdlib vars.
Because of a scoping bug in the trxl environment,
the variable name 'b' is leaked into the user env,
leading to crashes at least when map was called in
an already nested scope.

This patch renames the 'b' variable to '__body__',
therefore making it less likely that user defined
programs will clash.